+++ openstack/networking/v2/extensions/agents/results.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/gophercloud/gophercloud"
+	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/bgp/speakers"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/networks"
 	"github.com/gophercloud/gophercloud/pagination"
 )
@@ -172,4 +173,24 @@ func (r ListDHCPNetworksResult) Extract() ([]networks.Network, error) {
 
 	err := r.ExtractInto(&s)
 	return s.Networks, err
+}
+
+// ListBGPSpeakersResult is the respone of agents/{id}/bgp-speakers
+type ListBGPSpeakersResult struct {
+	pagination.SinglePageBase
+}
+
+func (r ListBGPSpeakersResult) IsEmpty() (bool, error) {
+	speakers, err := ExtractBGPSpeakers(r)
+	return 0 == len(speakers), err
+}
+
+// ExtractBGPSpeakers inteprets the ListBGPSpeakersResult into an array of BGP speakers
+func ExtractBGPSpeakers(r pagination.Page) ([]speakers.BGPSpeaker, error) {
+	var s struct {
+		Speakers []speakers.BGPSpeaker `json:"bgp_speakers"`
+	}
+
+	err := (r.(ListBGPSpeakersResult)).ExtractInto(&s)
+	return s.Speakers, err
 }